requirements.txt
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.1
MarkupSafe==2.1.1
pycparser==2.21
pydub==0.25.1
PyJWT==2.6.0
PyMySQL==1.0.2
pyrsistent==0.19.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
SpeechRecognition==3.8.1
urllib3==1.26.13
Werkzeug==2.2.2
gunicorn==20.0.4
beautifulsoup4==4.11.1
